修订测试代码

+++ tracker/radar_test.go
@@ -8,7 +8,7 @@ import (
 
 func TestConfig(t *testing.T) {
 	strategyCode := 82
-	rule := config.GetStrategyParameterByCode(strategyCode)
+	rule := config.GetStrategyParameterByCode(uint64(strategyCode))
 	fmt.Println(rule)
 	list := rule.StockList()
 	fmt.Println(list)
